Refactor sync status bar for re-use
