+++ android/quest/src/main/java/org/smartregister/fhircore/quest/ui/shared/components/SyncStatusView.kt
@@ -19,11 +19,15 @@ package org.smartregister.fhircore.quest.ui.shared.components
 import androidx.compose.animation.animateContentSize
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.Icon
 import androidx.compose.material.LinearProgressIndicator
 import androidx.compose.material.MaterialTheme
@@ -31,10 +35,21 @@ import androidx.compose.material.Text
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.CheckCircle
 import androidx.compose.material.icons.filled.Error
+import androidx.compose.material.icons.filled.KeyboardArrowDown
+import androidx.compose.material.icons.filled.KeyboardArrowUp
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.SideEffect
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
@@ -45,6 +60,9 @@ import com.google.android.fhir.sync.CurrentSyncJobStatus
 import com.google.android.fhir.sync.SyncJobStatus
 import com.google.android.fhir.sync.SyncOperation
 import java.time.OffsetDateTime
+import kotlin.time.Duration.Companion.seconds
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.launch
 import org.smartregister.fhircore.engine.ui.theme.AppTheme
 import org.smartregister.fhircore.engine.ui.theme.DangerColor
 import org.smartregister.fhircore.engine.ui.theme.DefaultColor
@@ -52,11 +70,139 @@ import org.smartregister.fhircore.engine.ui.theme.SubtitleTextColor
 import org.smartregister.fhircore.engine.ui.theme.SuccessColor
 import org.smartregister.fhircore.engine.ui.theme.SyncBarBackgroundColor
 import org.smartregister.fhircore.engine.util.annotation.PreviewWithBackgroundExcludeGenerated
-import org.smartregister.fhircore.quest.R
+import org.smartregister.fhircore.quest.ui.main.AppMainEvent
+import org.smartregister.fhircore.quest.ui.shared.models.AppDrawerUIState
 import org.smartregister.fhircore.quest.util.extensions.conditional
 
 const val TRANSPARENCY = 0.2f
 const val SYNC_PROGRESS_INDICATOR_TEST_TAG = "syncProgressIndicatorTestTag"
+
+@Composable
+fun SyncBottomBar(
+  isFirstTimeSync: Boolean,
+  appDrawerUIState: AppDrawerUIState,
+  onAppMainEvent: (AppMainEvent) -> Unit,
+  openDrawer: (Boolean) -> Unit,
+) {
+  val coroutineScope = rememberCoroutineScope()
+  val currentSyncJobStatus = appDrawerUIState.currentSyncJobStatus
+  val hideSyncCompleteStatus = remember { mutableStateOf(false) }
+  if (currentSyncJobStatus is CurrentSyncJobStatus.Succeeded) {
+    LaunchedEffect(Unit) {
+      coroutineScope.launch {
+        delay(7.seconds)
+        hideSyncCompleteStatus.value = true
+      }
+    }
+  }
+  val syncBackgroundColor =
+    when (currentSyncJobStatus) {
+      is CurrentSyncJobStatus.Failed -> DangerColor.copy(alpha = 0.2f)
+      is CurrentSyncJobStatus.Succeeded -> SuccessColor.copy(alpha = 0.2f)
+      is CurrentSyncJobStatus.Running -> SyncBarBackgroundColor
+      else -> Color.Transparent
+    }
+  var syncNotificationBarExpanded by remember { mutableStateOf(true) }
+  val bottomRadius =
+    if (!hideSyncCompleteStatus.value || currentSyncJobStatus is CurrentSyncJobStatus.Running) {
+      32.dp
+    } else 0.dp
+  val height =
+    when {
+      syncNotificationBarExpanded ->
+        if (currentSyncJobStatus is CurrentSyncJobStatus.Running) 114.dp else 80.dp
+      else -> 60.dp
+    }
+  if (
+    !isFirstTimeSync &&
+      currentSyncJobStatus != null &&
+      (currentSyncJobStatus is CurrentSyncJobStatus.Running ||
+        currentSyncJobStatus is CurrentSyncJobStatus.Failed ||
+        (!hideSyncCompleteStatus.value && currentSyncJobStatus is CurrentSyncJobStatus.Succeeded))
+  ) {
+    Box(
+      modifier =
+        Modifier.fillMaxWidth().animateContentSize().height(height).background(syncBackgroundColor),
+    ) {
+      Box(
+        modifier =
+          Modifier.fillMaxWidth()
+            .align(Alignment.TopStart)
+            .height(24.dp)
+            .clip(RoundedCornerShape(bottomStart = bottomRadius, bottomEnd = bottomRadius))
+            .background(Color.White),
+      ) {
+        Box(
+          modifier =
+            Modifier.align(Alignment.BottomStart)
+              .padding(start = 16.dp)
+              .height(20.dp)
+              .width(60.dp)
+              .clip(RoundedCornerShape(topStart = 8.dp, topEnd = 8.dp))
+              .background(syncBackgroundColor)
+              .clickable { syncNotificationBarExpanded = !syncNotificationBarExpanded },
+          contentAlignment = Alignment.Center,
+        ) {
+          Icon(
+            imageVector =
+              if (syncNotificationBarExpanded) {
+                Icons.Default.KeyboardArrowDown
+              } else Icons.Default.KeyboardArrowUp,
+            contentDescription = null,
+            tint =
+              when (currentSyncJobStatus) {
+                is CurrentSyncJobStatus.Failed -> DangerColor
+                is CurrentSyncJobStatus.Succeeded -> SuccessColor
+                else -> Color.White
+              },
+            modifier = Modifier.size(16.dp),
+          )
+        }
+      }
+      Box(
+        modifier = Modifier.align(Alignment.BottomStart).fillMaxWidth(),
+        contentAlignment = Alignment.Center,
+      ) {
+        val context = LocalContext.current
+        when (currentSyncJobStatus) {
+          is CurrentSyncJobStatus.Running -> {
+            SyncStatusView(
+              isSyncUpload = appDrawerUIState.isSyncUpload,
+              currentSyncJobStatus = currentSyncJobStatus,
+              minimized = !syncNotificationBarExpanded,
+              progressPercentage = appDrawerUIState.percentageProgress,
+              onCancel = { onAppMainEvent(AppMainEvent.CancelSyncData(context)) },
+            )
+            SideEffect { hideSyncCompleteStatus.value = false }
+          }
+          is CurrentSyncJobStatus.Failed -> {
+            SyncStatusView(
+              isSyncUpload = appDrawerUIState.isSyncUpload,
+              currentSyncJobStatus = currentSyncJobStatus,
+              minimized = !syncNotificationBarExpanded,
+              onRetry = {
+                openDrawer(false)
+                onAppMainEvent(AppMainEvent.SyncData(context))
+              },
+            )
+          }
+          is CurrentSyncJobStatus.Succeeded -> {
+            if (!hideSyncCompleteStatus.value) {
+              SyncStatusView(
+                isSyncUpload = appDrawerUIState.isSyncUpload,
+                currentSyncJobStatus = currentSyncJobStatus,
+                minimized = !syncNotificationBarExpanded,
+              )
+            }
+          }
+          else -> {
+            // No render required
+          }
+        }
+      }
+    }
+  }
+}
 
 @Composable
 fun SyncStatusView(
